Changed to pool page
