+++ src/pages/AddLiquidity/index.tsx
@@ -6,7 +6,7 @@ import { Plus } from 'react-feather'
 import ReactGA from 'react-ga4'
 import { RouteComponentProps } from 'react-router-dom'
 import { Text } from 'rebass'
-import styled, { ThemeContext } from 'styled-components'
+import { ThemeContext } from 'styled-components'
 import { ButtonError, ButtonLight, ButtonPrimary } from '../../components/Button'
 import { BlueCard, LightCard } from '../../components/Card'
 import { AutoColumn, ColumnCenter } from '../../components/Column'
@@ -15,8 +15,7 @@ import CurrencyInputPanel from '../../components/CurrencyInputPanel'
 import DoubleCurrencyLogo from '../../components/DoubleLogo'
 import { AddRemoveTabs } from '../../components/NavigationTabs'
 import Row, { RowBetween, RowFlat } from '../../components/Row'
-import { DFI, USDT } from '../../constants/index'
-import { MetaMaskInpageProvider } from '@metamask/providers'
+import { USDT } from '../../constants/index'
 import { PROXIES } from '../../constants'
 import { PairState, ProxyPair, usePairs2 } from '../../data/Reserves'
 import { useActiveWeb3React } from '../../hooks'
@@ -45,32 +44,6 @@ import { PoolPriceBar } from './PoolPriceBar'
 import { useIsTransactionUnsupported } from 'hooks/Trades'
 import { useProxies } from 'state/wallet/hooks'
 import { useTotalStake } from 'data/TotalSupply'
-
-const StyledBalanceMax = styled.button`
-  margin-top: 5px;
-  height: 30px;
-  background-color: ${({ theme }) => theme.primary5};
-  border: 1px solid ${({ theme }) => theme.primary5};
-  border-radius: 0.5rem;
-  font-size: 0.875rem;
-  text-decoration: underline;
-
-  font-weight: 500;
-  cursor: pointer;
-  margin-right: 0.5rem;
-  color: ${({ theme }) => theme.primaryText1};
-  :hover {
-    border: 1px solid ${({ theme }) => theme.primary4};
-  }
-  :focus {
-    border: 1px solid ${({ theme }) => theme.primary1};
-    outline: none;
-  }
-
-  ${({ theme }) => theme.mediaWidth.upToExtraSmall`
-    margin-right: 0.5rem;
-  `};
-`
 
 export default function AddLiquidity({
   match: {
@@ -355,27 +328,6 @@ export default function AddLiquidity({
       })
   }
 
-  async function addTokenFunction() {
-    const ethereum = window.ethereum as MetaMaskInpageProvider
-    try {
-      await ethereum.request({
-        method: 'wallet_watchAsset',
-        params: {
-          type: 'ERC20',
-          options: {
-            // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-            address: DFI[chainId!].address,
-            symbol: 'DFI',
-            decimals: 8,
-            image: 'https://s2.coinmarketcap.com/static/img/coins/64x64/5804.png'
-          }
-        }
-      })
-    } catch (error) {
-      console.log(error)
-    }
-  }
-
   const modalHeader = () => {
     return noLiquidity ? (
       <AutoColumn gap="25px">
@@ -645,7 +597,6 @@ export default function AddLiquidity({
           </AutoColumn>
         </Wrapper>
       </AppBody>
-      {account ? <StyledBalanceMax onClick={() => addTokenFunction()}>{'Add DFI to Metamask'}</StyledBalanceMax> : ' '}
     </>
   )
 }
